python_ai/requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
google-genai
langchain
langchain-core
langchain-google-genai
langchain-community
faiss-cpu #벡터DB
python-dotenv
httpx #Spring 내부 호출 시 사용